# Conflicts:
#	.idea/workspace.xml
#	examinationbackstage.iml
#	target/classes/application.yml
#	target/classes/com/bgs/examinationbackstage/ExaminationbackstageApplication.class
#	target/classes/com/bgs/examinationbackstage/config/SwaggerConfig.class
#	target/classes/com/bgs/examinationbackstage/controller/UserController.class
#	target/classes/com/bgs/examinationbackstage/controller/ZhuzhiController.class
#	target/classes/com/bgs/examinationbackstage/interceptor/DataBaseInterceptor.class
#	target/classes/com/bgs/examinationbackstage/mapper/UserMapper.class
#	target/classes/com/bgs/examinationbackstage/mapper/ZhuzhiMapper.class
#	target/classes/com/bgs/examinationbackstage/pojo/AccessTokenDto.class
#	target/classes/com/bgs/examinationbackstage/pojo/AuthTokenModel.class
#	target/classes/com/bgs/examinationbackstage/pojo/ExaminationAnswer.class
#	target/classes/com/bgs/examinationbackstage/pojo/ExaminationPaper.class
#	target/classes/com/bgs/examinationbackstage/pojo/Permission.class
#	target/classes/com/bgs/examinationbackstage/pojo/QuestionBank.class
#	target/classes/com/bgs/examinationbackstage/pojo/QuestionImg.class
#	target/classes/com/bgs/examinationbackstage/pojo/Role.class
#	target/classes/com/bgs/examinationbackstage/pojo/RolePermession.class
#	target/classes/com/bgs/examinationbackstage/pojo/TestPaperQuestions.class
#	target/classes/com/bgs/examinationbackstage/pojo/TestPaperUser.class
#	target/classes/com/bgs/examinationbackstage/pojo/User.class
#	target/classes/com/bgs/examinationbackstage/service/UserService.class
#	target/classes/com/bgs/examinationbackstage/service/ZhuzhiService.class
#	target/classes/com/bgs/examinationbackstage/service/serviceImp/UserServiceImpl.class
#	target/classes/com/bgs/examinationbackstage/service/serviceImp/ZhuzhiServiceImp.class
#	target/classes/com/bgs/examinationbackstage/utils/BaseResponse.class
#	target/classes/com/bgs/examinationbackstage/utils/Constant.class
#	target/classes/com/bgs/examinationbackstage/utils/EncryptUtil.class
#	target/classes/com/bgs/examinationbackstage/utils/SnowFlake.class
#	target/classes/com/bgs/examinationbackstage/utils/StatusCode.class
#	target/classes/mapper/UserMapper.xml
#	target/classes/mapper/ZhuzhiMapper.xml

+++ src/main/java/com/bgs/examinationbackstage/pojo/User.java
@@ -1,6 +1,7 @@
 package com.bgs.examinationbackstage.pojo;
 
 
+import io.swagger.models.auth.In;
 import lombok.Data;
 
 @Data
@@ -13,7 +14,7 @@ public class User {
   private long accountNumber;
   private String password;
   private String phone;
-  private long roleId;
+  private Integer roleId;
